- Support use of $VERSION and $BUILD replacement tokens in single-run action paths
- Refactor common code to shared static util class
- Improved error reporting

+++ Editor/FileUtility/FileOperation.cs
@@ -5,15 +5,10 @@ using UnityEditor;
 
 namespace SuperUnityBuild.BuildActions
 {
+    using Operation = FileUtility.Operation;
+
     public class FileOperation : BuildAction, IPreBuildAction, IPreBuildPerPlatformAction, IPostBuildAction, IPostBuildPerPlatformAction
     {
-        public enum Operation
-        {
-            Move,
-            Copy,
-            Delete,
-        }
-
         [BuildTool.FilePath(false)]
         public string inputPath;
         [BuildTool.FilePath(false)]
@@ -23,16 +18,19 @@ namespace SuperUnityBuild.BuildActions
 
         public override void Execute()
         {
+            string resolvedInputPath = FileUtility.ResolvePath(inputPath);
+            string resolvedOutputPath = FileUtility.ResolvePath(outputPath);
+
             switch (operation)
             {
                 case Operation.Copy:
-                    Copy(inputPath, outputPath);
+                    Copy(resolvedInputPath, resolvedOutputPath);
                     break;
                 case Operation.Move:
-                    Move(inputPath, outputPath);
+                    Move(resolvedInputPath, resolvedOutputPath);
                     break;
                 case Operation.Delete:
-                    Delete(inputPath);
+                    Delete(resolvedInputPath);
                     break;
             }
 
@@ -41,8 +39,8 @@ namespace SuperUnityBuild.BuildActions
 
         public override void PerBuildExecute(BuildReleaseType releaseType, BuildPlatform platform, BuildArchitecture architecture, BuildDistribution distribution, DateTime buildTime, ref BuildOptions options, string configKey, string buildPath)
         {
-            string resolvedInputPath = FolderOperation.ResolvePath(inputPath, releaseType, platform, architecture, distribution, buildTime, buildPath);
-            string resolvedOutputPath = FolderOperation.ResolvePath(outputPath, releaseType, platform, architecture, distribution, buildTime, buildPath);
+            string resolvedInputPath = FileUtility.ResolvePerBuildPath(inputPath, releaseType, platform, architecture, distribution, buildTime, buildPath);
+            string resolvedOutputPath = FileUtility.ResolvePerBuildPath(outputPath, releaseType, platform, architecture, distribution, buildTime, buildPath);
 
             switch (operation)
             {
@@ -62,7 +60,7 @@ namespace SuperUnityBuild.BuildActions
 
         protected override void DrawProperties(SerializedObject obj)
         {
-            bool containsWildcard = string.IsNullOrEmpty(inputPath) ? false : Path.GetFileNameWithoutExtension(inputPath).Contains("*");
+            bool containsWildcard = FileUtility.ContainsWildcard(inputPath);
 
             EditorGUILayout.PropertyField(obj.FindProperty("operation"));
 
@@ -77,21 +75,27 @@ namespace SuperUnityBuild.BuildActions
 
         private void Move(string inputPath, string outputPath, bool overwrite = true)
         {
-            bool containsWildcard = string.IsNullOrEmpty(inputPath) ? false : Path.GetFileNameWithoutExtension(inputPath).Contains("*");
+            bool success = true;
+            string errorString = "";
+            bool containsWildcard = FileUtility.ContainsWildcard(inputPath);
 
             if (!containsWildcard && !File.Exists(inputPath))
             {
                 // Error. Input does not exist.
-                return;
+                success = false;
+                errorString = $"Input \"{inputPath}\" does not exist.";
             }
 
-            if (!containsWildcard && overwrite && File.Exists(outputPath))
+            if (success && !containsWildcard && overwrite && File.Exists(outputPath))
             {
                 // Delete previous output.
-                FileUtil.DeleteFileOrDirectory(outputPath);
+                success = FileUtil.DeleteFileOrDirectory(outputPath);
+
+                if (!success)
+                    errorString = $"Could not overwrite existing file \"{outputPath}\".";
             }
 
-            if (containsWildcard)
+            if (success && containsWildcard)
             {
                 string inputDirectory = Path.GetDirectoryName(inputPath);
                 string outputDirectory = Path.GetDirectoryName(outputPath);
@@ -101,38 +105,58 @@ namespace SuperUnityBuild.BuildActions
 
                 for (int i = 0; i < fileList.Length; i++)
                 {
+                    if (!success)
+                        break;
+
                     string fileName = Path.GetFileName(fileList[i]);
                     string outputFile = Path.Combine(outputDirectory, fileName);
 
                     if (File.Exists(outputFile))
-                        FileUtil.DeleteFileOrDirectory(outputFile);
+                        success = FileUtil.DeleteFileOrDirectory(outputFile);
+
+                    if (!success)
+                        errorString = $"Could not overwrite existing file \"{outputPath}\".";
 
                     FileUtil.MoveFileOrDirectory(fileList[i], outputFile);
                 }
             }
-            else
+            else if (success)
             {
                 FileUtil.MoveFileOrDirectory(inputPath, outputPath);
+            }
+
+            if (!success && !string.IsNullOrEmpty(errorString))
+            {
+                BuildNotificationList.instance.AddNotification(new BuildNotification(
+                    BuildNotification.Category.Error,
+                    "File Move Failed.", errorString,
+                    true, null));
             }
         }
 
         private void Copy(string inputPath, string outputPath, bool overwrite = true)
         {
-            bool containsWildcard = string.IsNullOrEmpty(inputPath) ? false : Path.GetFileNameWithoutExtension(inputPath).Contains("*");
+            bool success = true;
+            string errorString = "";
+            bool containsWildcard = FileUtility.ContainsWildcard(inputPath);
 
             if (!containsWildcard && !File.Exists(inputPath))
             {
                 // Error. Input does not exist.
-                return;
+                success = false;
+                errorString = $"Input \"{inputPath}\" does not exist.";
             }
 
-            if (!containsWildcard && overwrite && File.Exists(outputPath))
+            if (success && !containsWildcard && overwrite && File.Exists(outputPath))
             {
                 // Delete previous output.
-                FileUtil.DeleteFileOrDirectory(outputPath);
+                success = FileUtil.DeleteFileOrDirectory(outputPath);
+
+                if (!success)
+                    errorString = $"Could not overwrite existing file \"{outputPath}\".";
             }
 
-            if (containsWildcard)
+            if (success && containsWildcard)
             {
                 string inputDirectory = Path.GetDirectoryName(inputPath);
                 string outputDirectory = Path.GetDirectoryName(outputPath);
@@ -142,28 +166,47 @@ namespace SuperUnityBuild.BuildActions
 
                 for (int i = 0; i < fileList.Length; i++)
                 {
+                    if (!success)
+                        break;
+
                     string fileName = Path.GetFileName(fileList[i]);
                     string outputFile = Path.Combine(outputDirectory, fileName);
 
                     if (File.Exists(outputFile))
-                        FileUtil.DeleteFileOrDirectory(outputFile);
+                        success = FileUtil.DeleteFileOrDirectory(outputFile);
+
+                    if (!success)
+                        errorString = $"Could not overwrite existing file \"{outputPath}\".";
 
                     FileUtil.CopyFileOrDirectory(fileList[i], outputFile);
                 }
             }
-            else
+            else if (success)
             {
                 FileUtil.CopyFileOrDirectory(inputPath, outputPath);
+            }
+
+            if (!success && !string.IsNullOrEmpty(errorString))
+            {
+                BuildNotificationList.instance.AddNotification(new BuildNotification(
+                    BuildNotification.Category.Error,
+                    "File Copy Failed.", errorString,
+                    true, null));
             }
         }
 
         private void Delete(string inputPath)
         {
-            bool containsWildcard = string.IsNullOrEmpty(inputPath) ? false : Path.GetFileNameWithoutExtension(inputPath).Contains("*");
+            bool success = true;
+            string errorString = "";
+            bool containsWildcard = FileUtility.ContainsWildcard(inputPath);
 
             if (!containsWildcard && File.Exists(inputPath))
             {
-                FileUtil.DeleteFileOrDirectory(inputPath);
+                success = FileUtil.DeleteFileOrDirectory(inputPath);
+
+                if (!success)
+                    errorString = $"Could not delete file \"{inputPath}\".";
             }
             else if (containsWildcard)
             {
@@ -174,12 +217,28 @@ namespace SuperUnityBuild.BuildActions
 
                 for (int i = 0; i < fileList.Length; i++)
                 {
-                    FileUtil.DeleteFileOrDirectory(fileList[i]);
+                    if (!success)
+                        break;
+
+                    success = FileUtil.DeleteFileOrDirectory(fileList[i]); ;
+
+                    if (!success)
+                        errorString = $"Could not delete file \"{fileList[i]}\".";
                 }
             }
             else
             {
                 // Error. File does not exist.
+                success = false;
+                errorString = $"Input \"{inputPath}\" does not exist.";
+            }
+
+            if (!success && !string.IsNullOrEmpty(errorString))
+            {
+                BuildNotificationList.instance.AddNotification(new BuildNotification(
+                    BuildNotification.Category.Error,
+                    "File Delete Failed.", errorString,
+                    true, null));
             }
         }
     }
